This commit contains manually cherry picked changes Niko made over the last ~15 months.
Going back to commit 120f9ba2740a09d8dd0d49b0f95a2f0227cceebb

Programs changed and added only affect the template matching workflow.

/programs/align_coordinates
/programs/make_particle_stack
/programs/niko_test.cpp
/programes/refine_template

There are several files where a confirmed float op was corrected:

IMAGE & refine2d
  Image::GetWeightedCorrelationWithImage fabs --> fabsf

Fix an incorrectly initialized counter, a long as 0.0 previously

  float Image::CosineMask
    long pixel counter = 0 (instead of 0.0) which seems like a correction

ctffind.cpp
  Fix unused bool and rename to avoid confustion (PowerSpectrumWasResampled)

find_dqe.cpp
  experimental changes, these are not necessarily validated. Refer to the last common ancester (niko) if needed to revert.

project3d
  The input pixel size now refers to the pixel size of the reference 3d as the output pixel size comes from the star file. This required inverting the scale factor.

refine3d
  set lower limit on number of results to refine to be one. Previously unspecified

refine_ctf
  addition of firstprivate in threading omp clause

programs/sharpen_map
  Just a return clause

Additionally, there is one change to the GpuImageClass cufftcallback for writing out half precision transforms for further statistical tracking. Fix
a mis-addressed pointer that is used when --disable-gpu-cache-hints is used at configure time.

+++ src/core/image.cpp
@@ -593,7 +593,7 @@ float Image::GetWeightedCorrelationWithImage(Image &projection_image, int *bins,
 			}
 			else
 			{
-				sum3 += fabs(cross_terms[i]);
+				sum3 += fabsf(cross_terms[i]);
 //				wxPrintf("i_u = %i, sum_a = %g, sum_b = %i, cross = %g\n", i, sum_a[i], sum_b[i], cross_terms[i]);
 //				r = fabsf(cross_terms[i] / sqrtf(sum_b[i]));
 			}
@@ -1843,7 +1843,6 @@ void Image::RotateQuadrants(Image &rotated_image, int quad_i)
 	MyDebugAssertTrue(logical_z_dimension == 1, "Error: attempting to rotate from 3D image");
 	MyDebugAssertTrue(rotated_image.is_in_memory, "Memory not allocated for receiving image");
 //	MyDebugAssertTrue(! is_in_real_space, "Image is in real space");
-	// These next two are redundant, right? Does it matter for the real space if it is square?
 	MyDebugAssertTrue(IsSquare(), "Image to rotate is not square");
 	MyDebugAssertTrue(rotated_image.logical_x_dimension == logical_x_dimension && rotated_image.logical_y_dimension == logical_y_dimension, "Error: Images different sizes");
 	MyDebugAssertTrue(quad_i == 0 || quad_i == 90 || quad_i == 180 || quad_i == 270, "Selected rotation invalid");
@@ -4017,7 +4016,7 @@ float Image::CosineMask(float wanted_mask_radius, float wanted_mask_edge, bool i
 			pixel_sum /= number_of_pixels;
 		}
 
-		pixel_counter = 0.0;
+		pixel_counter = 0;
 		for (k = 0; k < logical_z_dimension; k++)
 		{
 			z = powf(k - physical_address_of_box_center_z, 2);
